Updates for Python 3
This should work in Sublime Text 2 and 3 now.

The most significant change is removing the ElementTreeCDATA subclass,
which doesn't work with the changes to ElementTree in Python 3. There
may be ways to do the same thing with the static methods of etree, but
this seems too sketchy just to get a CDATA section. Instead, the XML
should be escaped like normal in the plain `content` tag.